% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{db_regions}
\alias{db_regions}
\title{Returns a list of regions from a database}
\usage{
db_regions(db)
}
\arguments{
\item{db}{Database from which regions should be listed}
}
\value{
Vector with regions in the database
}
\description{
Returns a list of regions from a database
}
\examples{

irena_db<-load_db(get_irena_db_files()[1])
db_regions(irena_db)
}
