% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{db_variables}
\alias{db_variables}
\title{Returns a list of variables from a database}
\usage{
db_variables(db)
}
\arguments{
\item{db}{Database from which variables should be listed}
}
\value{
Vector with regions in the database
}
\description{
Returns a list of variables from a database
}
\examples{
download_clean_save_irena()
irena_db<-load_db(get_irena_db_files()[1])
db_variables(irena_db)
}
