% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{download_clean_save_bp}
\alias{download_clean_save_bp}
\title{Download, clean and save BP World Energy Review Database to feather format}
\usage{
download_clean_save_bp(
 
    url = "https://www.bp.com/content/dam/bp/business-sites/en/global/corporate/xlsx/energy-economics/statistical-review/bp-stats-review-2019-all-data.xlsx"
)
}
\arguments{
\item{url}{Url to latest BP World Energy Review Database. Latest known is provided as default.}
}
\value{
Name of feather file were final cleaned database is stored
}
\description{
Download, clean and save BP World Energy Review Database to feather format
}
\examples{
download_clean_save_bp()
}
