% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{download_clean_save_irena}
\alias{download_clean_save_irena}
\title{Download, clean and save irena database to feather format}
\usage{
download_clean_save_irena(
 
    url = "https://www.irena.org/IRENADocuments/IRENA_RE_electricity_statistics_-_Query_tool.xlsm",
  target_dir = "./"
)
}
\arguments{
\item{url}{Url to latest IRENA Database. Latest known is provided as default.}
}
\value{
Name of feather file were final cleaned database is stored
#' @examples
download_clean_save_irena()
}
\description{
Download, clean and save irena database to feather format
}
