% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extract_trade_data_gas}
\alias{extract_trade_data_gas}
\title{Reads trade gas data from BP Database}
\usage{
extract_trade_data_gas(sheet)
}
\arguments{
\item{sheet}{Excel sheet with name "Gas - Inter-regional trade" from BP File}
}
\value{
A tibble with all trade values in the irenabpdata database format
}
\description{
Reads trade gas data from BP Database
}
