% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extract_trade_data_oil_coal}
\alias{extract_trade_data_oil_coal}
\title{Reads oil and gas trade data from BP Database}
\usage{
extract_trade_data_oil_coal(sheet, prefix = "Oil - ")
}
\arguments{
\item{sheet}{Excel sheet with name "Coal - Trade movements" or "Oil - Trade movements" from BP File}

\item{prefix}{Prefix to add to variable name. Preferably Oil or Coal, respectively.}
}
\value{
A tibble with all trade values in the irenabpdata database format
}
\description{
Reads oil and gas trade data from BP Database
}
