% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_bp_db_files}
\alias{get_bp_db_files}
\title{Returns a list of bp database files in the current directory}
\usage{
get_bp_db_files()
}
\value{
vector with bp database files
}
\description{
Returns a list of bp database files in the current directory
}
