% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_irena_db_files}
\alias{get_irena_db_files}
\title{Returns a list of irena database files in the current directory}
\usage{
get_irena_db_files()
}
\value{
vector with irena database files
}
\description{
Returns a list of irena database files in the current directory
}
