% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_per_capita_values}
\alias{get_per_capita_values}
\title{Returns per capita value of either IRENA or BP database}
\usage{
get_per_capita_values(db, db_type)
}
\arguments{
\item{db}{IRENA or BP Database}

\item{db_type}{Indicates if this is Irena or BP Database (Put "Irena" or "BP")}
}
\value{
A tibble with per capita values of variables in database
}
\description{
Returns per capita value of either IRENA or BP database
}
\examples{
download_clean_save_bp()
bp_db<-load_db(get_bp_db_files()[1])
get_per_capita_values(bp_db, "BP")
}
