% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_values_of_variable}
\alias{get_values_of_variable}
\title{Returns values of variables from db}
\usage{
get_values_of_variable(db, country, variable)
}
\arguments{
\item{db}{Database from which values should be extracted}

\item{country}{Country for which values should be extracted}

\item{variable}{Variable for which values should be extracted}
}
\value{
Vector with values in the database. Issues a warning if the number of rows in this table are 0.
}
\description{
Returns values of variables from db
}
\examples{
download_clean_save_bp()
bp_db<-load_db(get_bp_db_files()[1])
db_get_values_of_variables(bp_db, "Austria", "Oil - Refining capacity")
}
