% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{guess_url}
\alias{guess_url}
\title{Given a year, will guess the URL of the BP World Energy Review Excel File. Please observer
that BP mostly does not host prior versions of the review, so this will only work for the last 
available report year - also, it may fail, as BP frequently changes URLs.}
\usage{
guess_url(year)
}
\arguments{
\item{year}{Year for which to guess URL}
}
\value{
URL of BP Excel File
}
\description{
Given a year, will guess the URL of the BP World Energy Review Excel File. Please observer
that BP mostly does not host prior versions of the review, so this will only work for the last 
available report year - also, it may fail, as BP frequently changes URLs.
}
