% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{join_wb_db}
\alias{join_wb_db}
\title{Full join of worldbank database with either IRENA or BP database}
\usage{
join_wb_db(indicator, db, db_type)
}
\arguments{
\item{db}{IRENA or BP Database}

\item{db_type}{Indicates if this is Irena or BP Database (Put "Irena" or "BP")}

\item{wb_db}{Worldbank database}
}
\value{
A tibble with full_join of the two databases
}
\description{
Full join of worldbank database with either IRENA or BP database
}
\examples{
download_clean_save_bp()
bp_db<-load_db(get_bp_db_files()[1])
join_wb_db("NY.GDP.MKTP.PP.KD", bp_db, "BP")
}
