% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{load_latest_db_bp}
\alias{load_latest_db_bp}
\title{Loads the latest bp database in the current directory from a feather file}
\usage{
load_latest_db_bp()
}
\value{
A tibble with the database
#' @examples
download_clean_save_bp()
bp_db<-load_latest_db_bp()
}
\description{
Loads the latest bp database in the current directory from a feather file
}
