% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{load_latest_db_irena}
\alias{load_latest_db_irena}
\title{Loads the latest irena database in the current directory from a feather file}
\usage{
load_latest_db_irena()
}
\value{
A tibble with the database
#' @examples
download_clean_save_irena()
bp_db<-load_latest_irena_bp()
}
\description{
Loads the latest irena database in the current directory from a feather file
}
