% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_bp_primary_energy_mix}
\alias{plot_bp_primary_energy_mix}
\title{Plot primary energy supply per country from BP database}
\usage{
plot_bp_primary_energy_mix(bp, region)
}
\arguments{
\item{bp}{tibble with bp database}

\item{region}{Region to plot data}
}
\value{
A plot of the region
}
\description{
Plot primary energy supply per country from BP database
}
\examples{
download_clean_save_bp()
bp_db<-load_latest_db_bp()
plot_bp_primary_energy_mix(bp_db, "World")
}
