% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_bp_vs_irena}
\alias{plot_bp_vs_irena}
\title{Plots Irena vs BP data related to wind and solar capacities and generation}
\usage{
plot_bp_vs_irena(bp_irena_db)
}
\arguments{
\item{bp_irena_db}{Database that joins bp and irena data}
}
\value{
Scatterplot of variables
}
\description{
Plots Irena vs BP data related to wind and solar capacities and generation
}
\examples{
download_clean_save_irena()
download_clean_save_bp()
irena_db<-load_db(get_irena_db_files()[1])
bp_db<-load_db(get_bp_db_files()[1])
join_bp_irena<-join_bp_irena(bp_db, irena_db)
plot_bp_vs_irena(join_bp_irena)
}
