% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_irena_variables}
\alias{plot_irena_variables}
\title{plots irena variables}
\usage{
plot_irena_variables(
  irena_db,
  country,
  variable_filter = c(),
  indicator = "Generation"
)
}
\arguments{
\item{country}{The country}

\item{variable_filter}{Variables that should _not_ be shown}

\item{indicator}{Which indicator should be shown? Either "Generation" or "Capacity"}
}
\value{
an image object
}
\description{
plots irena variables
}
\examples{
download_clean_save_irena()
irena_db<-load_db(get_irena_db_files()[1])
plot_irena_variables(irena_db, "World")

irena_db<-load_db(get_irena_db_files()[1])
plot_irena_variables(irena_db, "World", variable_filter=FILTER_IRENA)

}
