% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{load_db}
\alias{load_db}
\title{loads one of the database from a feather file}
\usage{
load_db(filename_db)
}
\arguments{
\item{filename_db}{Filename of the database to load}
}
\value{
A tibble with the database
#' @examples
download_clean_save_irena()
irena_db<-load_db(get_irena_db_files()[1])
}
\description{
loads one of the database from a feather file
}
